requirements.txt
asgiref==3.5.0
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
bcrypt==3.2.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==37.0.1
defusedxml==0.7.1
Deprecated==1.2.13
distro==1.7.0
Django==4.0.4
django-cacheops==6.0
djangorestframework==3.13.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
et-xmlfile==1.1.0
funcy==1.17
httpie==3.1.0
idna==3.3
iniconfig==1.1.1
jsonschema==3.2.0
loadtest==0.0.2
multidict==6.0.2
openpyxl==3.0.9
packaging==21.3
paramiko==2.10.4
pluggy==1.0.0
py==1.11.0
pycparser==2.21
Pygments==2.12.0
PyNaCl==1.5.0
pyparsing==3.0.8
pyrsistent==0.18.1
PySocks==1.7.1
python-dotenv==0.20.0
pytz==2022.1
pywin32==227
PyYAML==5.4.1
redis==4.2.2
requests==2.27.1
requests-toolbelt==0.9.1
six==1.16.0
sqlparse==0.4.2
texttable==1.6.4
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.9
websocket-client==0.59.0
wrapt==1.14.0
